+++ src/App.jsx
@@ -1,16 +1,13 @@
-import logo from './logo.svg';
 import './assets/css/App.css';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import Home from './components/Home';
-import Inicio from './components/Inicio';
-import NavBar from './components/NavBar';
-import Nosotras from './components/Nosotras';
-import Contactanos from './components/Contactanos';
-
+import Home from './components/home/Home';
+import Inicio from './components/inicio/Inicio';
+import NavBar from './components/menu/NavBar';
+import Nosotras from './components/sobrenosotras/Nosotras';
+import Contactanos from './components/contacto/Contactanos';
 function App() {
   return (
     <div className="App">
-      
        <Router>
         <NavBar/>
         <Routes>
@@ -19,9 +16,8 @@ function App() {
           <Route path='/Nosotras' element={<Nosotras/>}/>
           <Route path='/Contactanos' element={<Contactanos/>}/>
         </Routes>
-       </Router>
+        </Router>
     </div>
   );
 }
-
 export default App;
