+++ src/assets/css/App.css
@@ -1,17 +1,14 @@
 .App {
   text-align: center;
 }
-
 body {
-    background-color: #5c6477;
+    background-color: #dee1e9;
 }
-
 .input-group{
   width:60%;
 }
-
 .displayFlex1 {
-  background-color: #1c1f63;
+  background-color: #e1e2ec;
   min-width: 50vh;
   display: flex;
   flex-direction: column;
@@ -19,18 +16,35 @@ body {
   font-size: 20px;
   color: rgb(231, 234, 232);
   border-radius: 40px;
-
 }
 .container2{
-  min-width: 100vh;
+  width: 400px;
   display: flex;
   flex-direction: column;
   align-items: center;
-  justify-content: space-between;
-  font-size: 30px;
-  color: rgb(15, 15, 15);
+  justify-content: center;
+  font-size: 20px;
+  color: rgb(246, 247, 243);
 }
 .listStyle{
   list-style: none;
 }
-
+.card{
+  background: rgb(14, 3, 53);
+  border:solid 4px rgb(204, 241, 38);
+  border-radius: 20px;
+}
+.container{
+  font-family: Impact, Haettenschweiler, 'Arial Narrow Bold', sans-serif;
+  color:rgb(20, 50, 79);
+  font-size: 20px;
+  text-align: center;
+  text-shadow: #dff30b;
+}
+.container3{
+  font-family: Impact, Haettenschweiler, 'Arial Narrow Bold', sans-serif;
+  color:rgb(8, 37, 65);
+  font-size: 30px;
+  text-align: center;
+  text-shadow: #0b1ef3;
+}
