--- src/components/Navbusqueda.js
@@ -1,11 +1,0 @@
-const Navbusqueda = () =>{
-return(
-    <nav className="Navbar bg-dark text-light mb-5">
-        <div className="container-fluid">
-            <h3 className="mx-auto"> ClimApp </h3>
-        </div>
-    </nav>
-);
-}
-
-export default Navbusqueda;
