--- src/components/Spin.js
@@ -1,13 +1,0 @@
-import React from "react";
-import  "../assets/css/Spin.css";
-const Spin = ()=>{
-    return(
-        <div className="lds-facebook">
-            <div></div>
-            <div></div>
-            <div></div>
-        </div>
-    );
-}
-
-export default Spin;
