+++ src/components/clima/CardForecast.jsx
@@ -1,14 +1,10 @@
 import React,{useState, useEffect} from 'react';
 import axios from 'axios';
-import '../assets/css/App.css';
-// import Spin from './Spin';
+import '../../assets/css/App.css';
 const CardForecast = (props)=> {
    const {ciudad} =props;
    const [data, setData] = useState({});
-   let leng="&lang=es";
-   let dias="&days=3";
-   let urlforecast =`https://api.weatherapi.com/v1/forecast.json?key=7b6c78c29c25447aaf5102820232205&q=`
-   let url=urlforecast + ciudad + dias + leng;
+   let url =`https://api.weatherapi.com/v1/forecast.json?key=7b6c78c29c25447aaf5102820232205&q=${ciudad}&days=3&lang=es`
     useEffect(() => {
       const buscaData = async () => {
         const { data } = await axios.get(url);
@@ -21,27 +17,30 @@ const CardForecast = (props)=> {
         
       }
       buscaData();
-    }, []);
+    }, [ciudad,url]);
 
     const { localidad, region, pais, pronostico }= data;
     return (
       <div className="container"> 
         {ciudad && (
-         <div className="row">
-            <div className="col displayFlex1 ">
+         <div className="row container3">
+            
               <p>{localidad}</p>
               <p>{region}</p>
               <p>{pais}</p>
-              {
+              <div className="row ">
+              { 
                   pronostico && pronostico.map(item=>
+                  <div className='card container2'>
                     <ul>
                     <li key={item.date} className="listStyle">
                       <p>{item.date}</p>
                       <p>Máx: {item.day.maxtemp_c} ºC</p>
                       <p>Mín: {item.day.mintemp_c} ºC</p>
-                      <img src={item.day.condition.icon}></img>
+                      <img src={item.day.condition.icon} width="80px" heigth="100" alt="Icono Tiempo"/>
                       <p>{item.day.condition.text}</p> </li>
                     </ul>
+                  </div>
                   )}
             </div>
           </div>
